Split Paths specs into separate file
